Generating fbc and fbp dynamically
Sorry for my ignorance.

I noticed that fbc and fbp are hard-coded.

Reading here https://developers.facebook.com/docs/marketing-api/conversions-api/parameters/fbp-and-fbc/

It seems these are supposed to be unique, first-party cookies. 

Is it necessary to generate these dynamically, or is hard-coding a value okay?

My understanding is that hard-coding these values would undermine the integrity of these values. 

Is that just because these values aren't needed for server-side conversions?

Thank you for your help, and sorry for my confusion.